+++ src/generated-types.ts
@@ -5877,47 +5877,45 @@ export interface components {
      * @description On-chain metadata stored in the minting transaction under label 721,
      * which adheres to https://cips.cardano.org/cips/cip25/
      */
-    asset_onchain_metadata_cip25:
-      | ({
-          /**
-           * @description Name of the asset
-           * @example My NFT token
-           */
-          name: string;
-          /**
-           * @description URI(s) of the associated asset
-           * @example ipfs://ipfs/QmfKyJ4tuvHowwKQCbCHj4L5T3fSj8cjs7Aau8V7BWv226
-           */
-          image: string | string[];
-          /**
-           * @description Additional description
-           * @example My NFT token description
-           */
-          description?: string | string[];
-          /**
-           * @description Mime sub type of image
-           * @example image/jpeg
-           */
-          mediaType?: string;
-          files?: ({
-            /**
-             * @description Name of the file
-             * @example myimage
-             */
-            name?: string;
-            /**
-             * @description Mime sub type of image
-             * @example image/jpeg
-             */
-            mediaType: string;
-            /**
-             * @description URI pointing to a resource of this mime type
-             * @example My NFT token description
-             */
-            src: string | string[];
-          } & { [key: string]: unknown })[];
-        } & { [key: string]: unknown })
-      | null;
+    asset_onchain_metadata_cip25: {
+      /**
+       * @description Name of the asset
+       * @example My NFT token
+       */
+      name: string;
+      /**
+       * @description URI(s) of the associated asset
+       * @example ipfs://ipfs/QmfKyJ4tuvHowwKQCbCHj4L5T3fSj8cjs7Aau8V7BWv226
+       */
+      image: string | string[];
+      /**
+       * @description Additional description
+       * @example My NFT token description
+       */
+      description?: string | string[];
+      /**
+       * @description Mime sub type of image
+       * @example image/jpeg
+       */
+      mediaType?: string;
+      files?: ({
+        /**
+         * @description Name of the file
+         * @example myimage
+         */
+        name?: string;
+        /**
+         * @description Mime sub type of image
+         * @example image/jpeg
+         */
+        mediaType: string;
+        /**
+         * @description URI pointing to a resource of this mime type
+         * @example My NFT token description
+         */
+        src: string | string[];
+      } & { [key: string]: unknown })[];
+    } & { [key: string]: unknown };
   };
   responses: {
     /** Bad request */