+++ src/functions/index.ts
@@ -1,19 +1,21 @@
 import fs from 'fs';
 import path from 'path';
 import YAML from 'yaml';
+import Ajv from 'ajv';
 
 import nutlinkAddressTickers from './custom-schemas/nutlink-address-tickers';
 import nutlinkTicker from './custom-schemas/nutlink-ticker';
 import scriptsJsonSchema from './custom-schemas/scripts-json';
 import txsMetadata from './custom-schemas/txs-metadata';
 
+const ajv = new Ajv({ strict: false });
 const file = fs.readFileSync(
   path.resolve(__dirname, '../../openapi.yaml'),
   'utf8',
 );
 const spec = YAML.parse(file);
 
-export default (endpointName: string) => {
+export const getSchemaForEndpoint = (endpointName: string) => {
   if (!spec.paths[endpointName]) {
     throw Error(
       `Missing Blockfrost OpenAPI schema for endpoint "${endpointName}".`,
@@ -149,4 +151,20 @@ export default (endpointName: string) => {
   // }
 
   return responses;
+};
+
+export const getSchema = (schemaName: string) => {
+  if (!spec.components.schemas[schemaName]) {
+    throw Error(`Missing Blockfrost OpenAPI schema with name "${schemaName}".`);
+  }
+
+  return spec.components.schemas[schemaName];
+};
+
+export const validateSchema = (schemaName: string, input: unknown) => {
+  const schema = getSchema(schemaName);
+  const validate = ajv.compile(schema);
+  const isValid = validate(input);
+
+  return { isValid, errors: validate.errors };
 };