+++ src/functions/schema.ts
@@ -3,10 +3,10 @@ import path from 'path';
 import YAML from 'yaml';
 import Ajv from 'ajv';
 
-import nutlinkAddressTickers from './custom-schemas/nutlink-address-tickers';
-import nutlinkTicker from './custom-schemas/nutlink-ticker';
-import scriptsJsonSchema from './custom-schemas/scripts-json';
-import txsMetadata from './custom-schemas/txs-metadata';
+import nutlinkAddressTickers from '../custom-schemas/nutlink-address-tickers';
+import nutlinkTicker from '../custom-schemas/nutlink-ticker';
+import scriptsJsonSchema from '../custom-schemas/scripts-json';
+import txsMetadata from '../custom-schemas/txs-metadata';
 
 const ajv = new Ajv({ strict: false });
 const file = fs.readFileSync(