+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project will adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) once it reaches maturity in v1.
 
+## [0.1.49] - 2022-11-26
+
+### Changed
+
+- update deps and yarn version
+
+### Added
+
+- test coverage
+- `getCIPstandard`, `getOnchainMetadataVersion`, `getOnchainMetadata` functions
+
 ## [0.1.48] - 2022-11-25
 
 ### Fixed